backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6

# Optional: Gemini (if you set GEMINI_API_KEY)
google-genai>=0.6.0

